clean up the module by removing unneeded models and classes
