--- imageviewer/src/main/java/com/giraffe/imageviewer/classifier/FakeClassifier.kt
@@ -1,9 +1,0 @@
-package com.giraffe.imageviewer.classifier
-
-import android.graphics.Bitmap
-
-class FakeClassifier: InappropriateImageClassifier {
-    override fun isInappropriate(bitmap: Bitmap): Boolean {
-        return true
-    }
-}
